+++ lib/responses/generate_pdf_from_documents_response.dart
@@ -1,0 +1,16 @@
+import 'package:pdf_combiner/responses/pdf_combiner_status.dart';
+
+class GeneratePdfFromDocumentsResponse {
+  String outputPath, message;
+  PdfCombinerStatus status;
+
+  GeneratePdfFromDocumentsResponse({
+    required this.status,
+    this.outputPath = "",
+    required this.message,
+  });
+
+  @override
+  String toString() =>
+      "GeneratePdfFromDocumentsResponse{outputPath: $outputPath, message: $message, status: $status }";
+}